Adjust routing and basic auth persistence
